+++ CHANGELOG.md
@@ -6,13 +6,15 @@
 
 ### Added
 
+- Добавлена валидация полученных ru-store-push токенов перед отправкой запроса.
 - Добавлена детализация исключений для ошибочных ответов RuStore: `InvalidArgumentException`, `NotFoundException`,
   `PermissionDeniedException`, `RuStoreInternalException`, `TooManyRequestsException`, `UnexpectedException`,
-  `MessageTooLargeException`,  `MessageTooLargeException`.
+  `MessageTooLargeException`, `InvalidPushTokenException`.
 - Добавлен `ResponseExceptionMapper`, преобразующий ошибочные HTTP-ответы RuStore в специализированные исключения.
 - Добавлена локальная проверка максимального объема сообщения RuStore перед отправкой HTTP-запроса.
+- Добавлена локальная проверка ru-store-push токенов: токен должен быть непустой строкой не больше 64 байт.
+- Добавлен `RuStoreMessageValidator`, отвечающий за локальную валидацию ru-store-push токенов и размера сообщения.
 - Класс `RuStoreReport` дополнен методом `target()`, возвращающим ru-store-push токен, для которого сформирован отчет.
-- Добавлена проверка максимального объем сообщения (4096 байт) перед отправкой. 
 
 ### Changed